+++ plasma-engine/src/diff.rs
@@ -178,6 +178,7 @@ mod tests {
                 .map(|(k, v)| (k.to_string(), v.to_string()))
                 .collect::<BTreeMap<_, _>>(),
             git: Default::default(),
+            file_contents: BTreeMap::new(),
         }
     }